Using touch input to hit strep.
Only works for centre strep. On hit, it changes to bicillin.
When cycle loops, it is a strep again.

Lot's more work to do, just proof of concept still.

Well done to #FightTheFever, winners of #hacknz2015.
The game concept was a factor.
This game demo was only the tiniest part of their success.